refactor: 태그 상단 컨테이너 분리
