+++ src/components/search/index.tsx
@@ -1,0 +1,9 @@
+import Map from './Map';
+
+function Index() {
+  return (
+    <Map />
+  );
+}
+
+export default Index;
